services/llm-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.0.1
openai==1.78.1
pydantic==2.11.4

